Version unchanged
Only thing I managed to do is a rework of the collision system (which took me 3 whole hours to debug, just because I put a line in the wrong place) and kinda fixed the arrow sprite bug, but now the arrow sprites don't disapper unless you get a new hand.

No changelog for now, pretty useless only for that. I'll include those in the version change.